YalseeSession with Redis and local fallback storage is done (#636)

+++ src/main/java/io/kyberorg/yalsee/services/YalseeSessionService.java
@@ -1,24 +1,19 @@
 package io.kyberorg.yalsee.services;
 
-import io.kyberorg.yalsee.constants.App;
 import io.kyberorg.yalsee.events.YalseeSessionUpdatedEvent;
 import io.kyberorg.yalsee.models.dao.YalseeSessionLocalDao;
 import io.kyberorg.yalsee.models.dao.YalseeSessionRedisDao;
 import io.kyberorg.yalsee.models.redis.YalseeSession;
-import io.kyberorg.yalsee.result.OperationResult;
 import io.kyberorg.yalsee.session.Device;
 import io.kyberorg.yalsee.utils.AppUtils;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.StringUtils;
 import org.greenrobot.eventbus.EventBus;
 import org.greenrobot.eventbus.Subscribe;
 import org.springframework.stereotype.Service;
 
 import javax.annotation.PostConstruct;
 import javax.annotation.PreDestroy;
-import javax.servlet.http.Cookie;
-import java.util.Objects;
 import java.util.Optional;
 
 @Slf4j
@@ -26,24 +21,22 @@ import java.util.Optional;
 @Service
 public class YalseeSessionService {
     private static final String TAG = "[" + YalseeSessionService.class.getSimpleName() + "]";
-    private static final String ERR_DEVICE_OBJECT_IS_NULL = "Device object is null";
-    private static final String ERR_COOKIE_OBJECT_IS_NULL = "Cookie object is null";
-    private static final String ERR_COOKIE_VALUE_IS_EMPTY = "Cookie Value is empty";
-    private static final String ERR_MALFORMED_COOKIE_VALUE = "Cookie value length mismatch";
-    private static final String ERR_RECORD_NOT_FOUND = "No corresponding record found in Redis";
-    private static final String ERR_HACK_ATTEMPT = "Record with ID exists, but info not match. Got Stolen Cookie.";
-    private static final String ERR_SESSION_EXPIRED = "Session expired";
 
     private final YalseeSessionRedisDao userSessionDao;
     private final YalseeSessionLocalDao fallbackLocalDao;
-    private final AppUtils appUtils;
-
 
     @PostConstruct
     public void init() {
         EventBus.getDefault().register(this);
     }
 
+    /**
+     * Creates new {@link YalseeSession}.
+     *
+     * @param device linked {@link Device} session created from.
+     * @return brand-new session with filled in {@link Device} info.
+     * @throws IllegalArgumentException when {@link Device} object is {@code null}.
+     */
     public YalseeSession createNew(final Device device) {
         if (device == null) throw new IllegalArgumentException("Device cannot be null");
         YalseeSession newSession = new YalseeSession();
@@ -60,6 +53,7 @@ public class YalseeSessionService {
      * Stores {@link YalseeSession} to Redis.
      *
      * @param session object to save
+     * @throws IllegalArgumentException if {@link YalseeSession} is {@code null}.
      */
     public void save(final YalseeSession session) {
         if (session == null) throw new IllegalArgumentException("Session cannot be null");
@@ -88,10 +82,17 @@ public class YalseeSessionService {
         return yalseeSession;
     }
 
+    /**
+     * Removes session from storages (local and remote), if it exists.
+     *
+     * @param session {@link YalseeSession} object to delete.
+     * @throws IllegalArgumentException if {@link YalseeSession} is {@code null}.
+     */
     public void destroySession(final YalseeSession session) {
+        if (session == null) throw new IllegalArgumentException("Session cannot be null");
         try {
-            userSessionDao.delete(session.getSessionId());
             fallbackLocalDao.delete(session.getSessionId());
+            userSessionDao.delete(session.getSessionId());
             log.info("{} {} {} destroyed", TAG, YalseeSession.class.getSimpleName(), session.getSessionId());
         } catch (Exception e) {
             log.error("{} failed to delete user session. Session ID: {}. Reason: {}",
@@ -107,59 +108,6 @@ public class YalseeSessionService {
             save(event.getSession());
             log.debug("{} Session '{}' updated", TAG, event.getSession().getSessionId());
         }
-    }
-
-    public Cookie createCookie(final YalseeSession us) {
-        Cookie cookie = new Cookie(App.CookieNames.USER_SESSION_COOKIE, us.getSessionId());
-        cookie.setMaxAge(appUtils.getSessionTimeout());
-        cookie.setSecure(us.getDevice().getWebBrowser().isSecureConnection());
-        cookie.setHttpOnly(true);
-        return cookie;
-    }
-
-    public OperationResult checkCookie(Cookie cookie, Device currentDevice) {
-        //Device null check
-        if (Objects.isNull(currentDevice)) {
-            return OperationResult.generalFail().withMessage(ERR_DEVICE_OBJECT_IS_NULL);
-        }
-        //Cookie Value aka Session Id
-        if (Objects.isNull(cookie)) {
-            return OperationResult.malformedInput().withMessage(ERR_COOKIE_OBJECT_IS_NULL);
-        }
-        String cookieValue = cookie.getValue();
-        if (StringUtils.isBlank(cookieValue)) {
-            return OperationResult.malformedInput().withMessage(ERR_COOKIE_VALUE_IS_EMPTY);
-        }
-        if (cookieValue.length() != YalseeSession.SESSION_ID_LEN) {
-            return OperationResult.malformedInput().withMessage(ERR_MALFORMED_COOKIE_VALUE);
-        }
-
-        //lookup by sessionId
-        Optional<YalseeSession> sessionRecord = this.getSession(cookieValue);
-        if (sessionRecord.isEmpty()) {
-            return OperationResult.elementNotFound().withMessage(ERR_RECORD_NOT_FOUND);
-        }
-        YalseeSession session = sessionRecord.get();
-        //valid (not expired)
-        if (session.expired()) {
-            invalidateAffectedSession(session);
-            return OperationResult.generalFail().withMessage(ERR_SESSION_EXPIRED);
-        }
-        //compare devices
-        Device savedDevice = session.getDevice();
-        if (!currentDevice.isSameDevice(savedDevice)) {
-            //device mismatch - stolen cookie
-            this.invalidateAffectedSession(session);
-            return OperationResult.banned().withMessage(ERR_HACK_ATTEMPT);
-        }
-
-        //all good
-        return OperationResult.success().addPayload(session);
-    }
-
-    public void invalidateAffectedSession(final YalseeSession session) {
-        log.info("{} deleting affected session {}", TAG, session.getSessionId());
-        this.destroySession(session);
     }
 
     @PreDestroy
